refactor: Merge fightstick files
Remove separation between fightstick.py and fightstick_hb.py. We can address redundancy later.
